Display H, S, L with 1 decimal instead of 2
The sliders are too coarse for the hundredths digits to matter. For example, a one-pixel drag on Hue changes from it 181.59 to 182.65. The extra digits at the end just make the numbers look more confusing – 181.6 and 182.6 are precise enough.

With experiments, I determined that changing the value of a parameter by 0.1 never changes a R, G, or B byte by more than one. It also looks like all 24-bit colors can be represented with only tenths digits for each of the values – extra digits are not needed to uniquely specify a color.

If the user really wants two decimal points of precision, they can just type the extra digits in the H, S, and L fields.

Since the maximum width of the text is less now, I shrunk the text boxes towards the sliders, increasing the margin between the sliders and the swatch.