Fix red outline on Firefox due to input validation
The three H, S, L text fields in the color picker had red outlines in Firefox whenever they contained a non-integral value such as “57.8”, even when such values were automatically input when dragging the slider. This was because the `<input>`’s default step value of 1 also implied that only multiples of 1 were acceptable, and when the text fields failed validation, Firefox drew a red outline around them. The new step value of `any` means that any amount of precision is acceptable.

I didn’t completely disable the red outlines, which is possible with CSS, because they are still helpful if the user types a non-number into one of the fields.