Fix parsing of inputted hex colors
Before this fix, if you pasted a color “aa22cc” with no leading “#”, the picker accepted the color as correct but skipped its first character, interpreting the color as #a22cc0. Now it will accept the color as correct and interpret it as #aa22cc.

It’s not important, but I will mention that the picker will do nothing if there is extraneous text around the inputted hex code, e.g. if the input is “I can’t decide between 000000 and fe57a1”. I could have instead made it search for the first hex-code-like string within the input. I don’t have strong feelings on which behavior is better, so I implemented the first one.